feat(myrag): wizard 5-step source-first + BaseConnector + SourceConfig
Wizard redesigned to start with source selection:
1. Source: Legifrance, Fichier, Repertoire, Drive, Nextcloud, Resana
2. Identification: name, prompt (pre-selected by source), strategy, contact
3. Donnees: upload/import adapted to source type
4. Evaluation: playground + Q&R dataset (optional)
5. Publication: modes, visibility, widget snippet

Backend:
- SourceConfig dataclass (type, url, auth, refresh)
- BaseConnector interface (list_documents, fetch_document, check_updates)
- connectors/ directory for pluggable source implementations

Frontend:
- Step 1: source cards with icons, refresh badge, suggested strategy
- Step 2: identification with pre-filled defaults based on source
- Steps 3-5: renumbered with updated navigation links
- WizardStepper: 5 steps

TDD: 135 tests passing
