+++ integrations/myrag/frontend/pages/admin/create/index.vue
@@ -8,133 +8,97 @@
     </nav>
 
     <h1 class="fr-h3">Creer une collection</h1>
-    <p class="fr-text--lg fr-mb-4w">Etape 1 sur 4 — Identification generale</p>
+    <p class="fr-text--lg fr-mb-4w">Etape 1 sur 5 — D'ou viennent vos documents ?</p>
 
     <WizardStepper :current-step="1" />
 
-    <!-- Step 1: Identification -->
-    <div class="fr-col-8">
-      <div class="fr-input-group">
-        <label class="fr-label" for="name">Nom de la collection *</label>
-        <input id="name" class="fr-input" v-model="form.name" placeholder="ceseda-v4" />
-      </div>
-
-      <div class="fr-input-group fr-mt-2w">
-        <label class="fr-label" for="desc">Description</label>
-        <textarea id="desc" class="fr-input" v-model="form.description" rows="2"
-               placeholder="Code de l'entree et du sejour des etrangers"></textarea>
-      </div>
-
-      <div class="fr-select-group fr-mt-2w">
-        <label class="fr-label">Prompt systeme de la collection</label>
-        <select class="fr-select" v-model="form.prompt_template">
-          <option v-for="tpl in templates" :key="tpl.key" :value="tpl.key">
-            {{ tpl.icon }} {{ tpl.name }}
-          </option>
-        </select>
-      </div>
-
-      <div class="fr-select-group fr-mt-2w">
-        <label class="fr-label">Strategie de decoupage</label>
-        <select class="fr-select" v-model="form.strategy">
-          <option value="auto">Automatique (detection du type)</option>
-          <option value="article">Par article (code juridique)</option>
-          <option value="section">Par section (rapport)</option>
-          <option value="qr">Par Q&R (FAQ)</option>
-          <option value="length">Par longueur fixe</option>
-        </select>
-      </div>
-
-      <div class="fr-grid-row fr-grid-row--gutters fr-mt-2w">
-        <div class="fr-col-6">
-          <div class="fr-select-group">
-            <label class="fr-label">Sensibilite</label>
-            <select class="fr-select" v-model="form.sensitivity">
-              <option value="public">Public</option>
-              <option value="internal">Interne</option>
-              <option value="restricted">Restreint</option>
-              <option value="confidential">Confidentiel</option>
-            </select>
+    <div class="myrag-card-grid">
+      <div v-for="src in sources" :key="src.key"
+           :class="['fr-card', 'fr-enlarge-link', { 'fr-card--shadow': selected === src.key }]"
+           @click="selected = src.key"
+           style="cursor:pointer;">
+        <div class="fr-card__body">
+          <div class="fr-card__content">
+            <h3 class="fr-card__title">
+              <span>{{ src.icon }} {{ src.name }}</span>
+            </h3>
+            <p class="fr-card__desc">{{ src.description }}</p>
+            <div class="fr-card__start">
+              <span v-if="src.refresh" class="fr-badge fr-badge--sm fr-badge--new">Refresh auto</span>
+              <span class="fr-badge fr-badge--sm fr-badge--info">{{ src.strategy }}</span>
+            </div>
           </div>
         </div>
-        <div class="fr-col-6">
-          <div class="fr-select-group">
-            <label class="fr-label">Portee</label>
-            <select class="fr-select" v-model="form.scope">
-              <option value="public">General (tous)</option>
-              <option value="group">Groupe</option>
-              <option value="private">Prive</option>
-            </select>
-          </div>
+        <div v-if="selected === src.key" class="fr-card__header" style="background:#000091;padding:4px;text-align:center;">
+          <span style="color:white;font-size:0.8rem;font-weight:bold;">✓ Selectionne</span>
         </div>
       </div>
+    </div>
 
-      <fieldset class="fr-fieldset fr-mt-3w">
-        <legend class="fr-fieldset__legend">Options</legend>
-        <div class="fr-fieldset__element">
-          <div class="fr-checkbox-group">
-            <input type="checkbox" id="graph" v-model="form.graph_enabled" />
-            <label class="fr-label" for="graph">Activer le graph de references</label>
-          </div>
-        </div>
-        <div class="fr-fieldset__element">
-          <div class="fr-checkbox-group">
-            <input type="checkbox" id="ai_summary" v-model="form.ai_summary_enabled" />
-            <label class="fr-label" for="ai_summary">Resume IA pour les articles longs</label>
-          </div>
-        </div>
-      </fieldset>
-
-      <!-- Contact info -->
-      <div class="fr-input-group fr-mt-3w">
-        <label class="fr-label">Responsable / contact</label>
-        <input class="fr-input" v-model="form.contact_name" placeholder="Nom du responsable" />
-      </div>
-      <div class="fr-input-group fr-mt-1w">
-        <label class="fr-label">Email de contact</label>
-        <input class="fr-input" type="email" v-model="form.contact_email" placeholder="responsable@example.com" />
-      </div>
-
-      <div class="fr-btns-group fr-btns-group--right fr-mt-4w">
-        <button class="fr-btn" @click="createAndNext" :disabled="creating || !form.name.trim()">
-          {{ creating ? 'Creation...' : 'Suivant →' }}
-        </button>
-      </div>
-
-      <div v-if="error" class="fr-alert fr-alert--error fr-mt-2w"><p>{{ error }}</p></div>
+    <div class="fr-btns-group fr-btns-group--right fr-mt-4w">
+      <button class="fr-btn" @click="next" :disabled="!selected">
+        Suivant →
+      </button>
     </div>
   </div>
 </template>
 
 <script setup lang="ts">
 const router = useRouter()
-const { get, post } = useApi()
+const selected = ref('')
 
-const form = ref({
-  name: '', description: '', strategy: 'auto', sensitivity: 'public',
-  prompt_template: 'generic', scope: 'group', graph_enabled: false,
-  ai_summary_enabled: false, contact_name: '', contact_email: '',
-})
-const templates = ref<any[]>([])
-const creating = ref(false)
-const error = ref('')
+const sources = [
+  {
+    key: 'legifrance',
+    icon: '⚖️',
+    name: 'Legifrance',
+    description: 'Code, loi, ordonnance depuis l\'API PISTE. Decoupage automatique par article avec hierarchie.',
+    refresh: true,
+    strategy: 'article',
+  },
+  {
+    key: 'file',
+    icon: '📄',
+    name: 'Fichier unique',
+    description: 'Upload un fichier (PDF, MD, TXT, DOCX, images, audio). Detection automatique du type.',
+    refresh: false,
+    strategy: 'auto',
+  },
+  {
+    key: 'directory',
+    icon: '📁',
+    name: 'Repertoire local',
+    description: 'Upload un dossier ou un ZIP contenant plusieurs fichiers a indexer ensemble.',
+    refresh: false,
+    strategy: 'auto',
+  },
+  {
+    key: 'drive',
+    icon: '☁️',
+    name: 'Suite Numerique Drive',
+    description: 'Connecter un dossier Drive (Suite Numerique). Synchronisation automatique des modifications.',
+    refresh: true,
+    strategy: 'auto',
+  },
+  {
+    key: 'nextcloud',
+    icon: '📦',
+    name: 'Nextcloud',
+    description: 'Connecter un dossier Nextcloud via l\'API OCS. Synchronisation automatique.',
+    refresh: true,
+    strategy: 'auto',
+  },
+  {
+    key: 'resana',
+    icon: '🗂️',
+    name: 'Resana',
+    description: 'Connecter un espace Resana. Synchronisation automatique des documents.',
+    refresh: true,
+    strategy: 'auto',
+  },
+]
 
-async function createAndNext() {
-  creating.value = true
-  error.value = ''
-  try {
-    await post('/api/collections', form.value)
-    router.push(`/admin/create/step-2?collection=${form.value.name}`)
-  } catch (e: any) {
-    error.value = e.message
-  }
-  creating.value = false
+function next() {
+  router.push(`/admin/create/step-2?source=${selected.value}`)
 }
-
-onMounted(async () => {
-  try {
-    const data = await get('/api/collections/templates')
-    templates.value = data.templates || []
-  } catch (e) {}
-})
 </script>
